SNAT Ext. Router scheduling broken w/ upgrade
SNAT requires that all the nodes be in the operational datastore
at the time the external router configurations are pushed to ODL.
At this point the external router is "scheduled", that is, a
switch is chosen to get the flows and track the sessions. In the
case of upgrade, there are in fact no nodes connected to ODL.

Solution:
Wait until all the nodes are connected and only then schedule
the external routers.
1) if the upgrading flag is set, don't schedule external routers
2) When the listener fires, schedule all the external routers

Change-Id: Ib349450905ef9c33fe3640753bb88e65c8ebf967
Signed-off-by: Josh <jhershbe@redhat.com>