Upgrade: Remote FIB entry missing for ext. gateway
Root cause of this is that the ARP to discover the
external gateway's mac is triggered at router/subnet
creation time and the code assumes the ovsdb models
are in place. However, in the case of upgrade they are
not yet present as the switches have not yet reconnected.

Fix: At the point where the upgrading flag is unset we
know that all switches have reconnected. Listen for that
transition and trigger the ARP for all routers.

Change-Id: I6cc6045937e2d9a936596e1a8bc36b7a0cbb70e3
Signed-off-by: Josh <jhershbe@redhat.com>