Upgrade+conntrack, default FIB flow missing
N.B.: Depends on https://git.opendaylight.org/gerrit/#/c/65894/

This patch contains fixes for the two problems that
cause this issue:

1) The default fib route can not be configured before
the VpnInstanceOpDataEntry's dpn<->vpn mapping exists.
I've added code that waits for that to show up.

2) The default fib route's action is to send to the
external network group which must be configured on the
switch before the flow is configured. This issue was resolved
here [0] for controller based snat but an attempted
identical switch for conntrack based snat did not work.
As such, I implemented a (pretty ugly hack) where the
code literally waits for the group to show up in the
node's config md-sal. I am very open to any better ideas.

[0] https://git.opendaylight.org/gerrit/#/c/64983/

Change-Id: I7a4e75b04b171b5a6c80c74e75a410312ef3dee6
Signed-off-by: Josh <jhershbe@redhat.com>