Update patch set 8
Patch Set 8:

It seems the issue that this patch is attempting to address is the following:

  - a data delete is queued in the batching manager
  - before the delete is executed, another component wants to re-add that data 
    and issues a read it get he existing data
  - however the data still exists in the DS thus the add is ignored

Assuming that all is correct then it seems to me that the batching mechanism is inherently flawed wrt reads. Reads should be batched along with writes and deletes. This should be handled transparently by the BatchedTransaction et al.

Patch-set: 8