Update patch set 11
Patch Set 11:

It seems you moved the maps from BatchedTransaction to DataUpdates. I don't understand the purpose of DataUpdates.  From the patch summary, it seems the only issue is what I had commented so, with the read method added to the BatchedResourceManager, the only change that should be needed in this patch is to change BatchedTransaction to use it.

Patch-set: 11