Support directional-dpls & cleanup sfc-translator
This patch adds support for the new directional-dpl sfc spec. It allows
to setup both the egress and ingress ports of a port pair as distinct
service funcion data plane locators so that they are used appropriately
depending on the direction of a symmetric path.

This patch also fully transitions to the logical sff configuration model
cleaning up code that applied to the old configuration model.

Change-Id: Ibee6e2707878932ae1efbfd4ab0eae8a78056437
Signed-off-by: Jaime Caamaño Ruiz <jcaamano@suse.com>