Update patch set 1
Patch Set 1: Code-Review+1

That's the right thing to do, see https://git.opendaylight.org/gerrit/#/c/65805/. Not sure if to merge this right now or wait for the overall mess to first get stabilitzed? Everything is in a mess and broken right now, and my personal view is that a big revert is one of the options that must be considered; I can't remememer if the new method with the class was already available before yangtools 2.0.0, which if not would make a revert harder.

Patch-set: 1
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1