implement scale in rpc
implemented the following rpcs
Future<RpcResult<Void>> scaleinComputesStart(ScaleinComputesStartInput input)

This rpc signals start of scaling in operation of the compute nodes.
This rpc marks the nodes as tombstoned in the config topology tree
by setting the bridge external id TOMBSTONED for the given node

Future<RpcResult<Void>> scaleinComputesRecover(ScaleinComputesRecoverInput input)

This rpc signals scale in operation failed in the middle and the given nodes
needs to be recovered.
This rpc unmarks the nodes as tombstoned in the config topology tree

Future<RpcResult<Void>> scaleinComputesEnd(ScaleinComputesEndInput input)
This signals end of the scale in operation for these nodes
This rpc deletes the nodes from config topology

Added the following scale in utility methods

boolean isDpnTombstoned(BigInteger dpnId)

Applications can check this api to see if the dpn is in the process of
scale in operation.

void addOnRecoveryCallback(Function<BigInteger, Void> callback)

Applications can add the recovery callback to get notification
on the recovered dpns so that they can take action on them by 
potentially choosing them as designated for its particular service

List<BigInteger> filterTombStoned(List<BigInteger> dpns)

Returns a list of valid dpns which are not part of scaled in dpn list
This api can be used by applications to get the candidate list for
designated dpns

Change-Id: I474f1cd2c1cefaea00f9d04eb871cc216dc84f02
Signed-off-by: xcheara <chetan.arakere@altencalsoftlabs.com>
Signed-off-by: Sam Hague <shague@redhat.com>
Signed-off-by: xcheara <chetan.arakere@altencalsoftlabs.com>