Scale-in Support for SNAT NAPT Switch
Description
-----------

1) New Listener - NatScaleIn added to handle remove event which is
registered on bridge-external-ids(network-topology) as a callboack.
Whenever Scale-in of an particular DPN fails and TOMSTONE=true entry been
rermoved then, if that DPN have some
Router Presence(with SNAT enabled) and if the correspodning NAPT switch is
missing, this DPN will be re-elected as NAPT switch.

2) New Listener - NatTepChangeLister is added to handle remove event. The
Existing logic of NatTunnelInterfaceStateListener.remove() been moved into
this Listener. Whenever the TEP on a given DPNID is removed, this API take
care of withdrawing the FIB entries for those Floating-IP existing on this
DPN and perform re-election of NAPT Switch for a VRF to which the current
DPN is elected as NAPT Switch.

3) Changes done to to validate the candidate list of DPNs set with
TOMSTONE=true and exclude them during NAPT Switch election.

4) Change done to remove installation of (47->21) flow as part of single
transcation. When Multiple router based of same BGPVPN elects same DPN as
NAPT switch, then both try to install same flow simultanously resulting in
Lock Exception. Change done to install this flow individually and not part
of single writeTransaction.

Change-Id: I992a25aa028ca49a730ee2813c6d0de5cbe84b82
Signed-off-by: xcheara <chetan.arakere@altencalsoftlabs.com>