NETVIRT-1065 Handle new bridges
If a provider network's value is a bridge name we
patch it to br-int. If the bridge does not exist prior
to the discovery of the provider mapping ODL will assume
the provider net is a local port and add it to br-int.
However, a user can remove that port and then add the bridge
and now the code will detect that addition and create the
patch ports.

Change-Id: Ieae1a37ccf90caa51566864be14d840319429e61
Signed-off-by: Josh <jhershbe@redhat.com>
(cherry picked from commit 216a424f69d254cbafe86a06cd90a42ed7fb6d2e)