Support symmetric chain classification
Symmetric chain classification through networking-sfc requires a flow
classifier that defines both a source and a destination logical port.
This patch adds support for such scenario.

RedirectToSfc can now specify a SFP name. Classification for the
forward RSP of that SFP will be done on the source port. Classification
for the reverse RSP of that SFP will be done on the destination port.
Sfc translator has been modified to specify a SFP name in
RedirectToSfc instead of a RSP name.

When RedirectToSfc specifies a RSP name, if the RSP is not reverse,
classification will be done in the source port. Classification will
happen in the destination port otherwise. No classification will be
done if no source port is specified for a reverse RSP, and viceversa.
If a network is specified instead, all the ports of the network will be
considered source ports if the RSP is not reverse, and destination
ports otherwise, meaning that classification will happen on those ports
wether the RSP is forward or reverse. This maintains legacy behavior
with respect network classification.

Change-Id: I86d85a58688367591784f4361aacd095e9cbccf0
Signed-off-by: Jaime Caamaño Ruiz <jcaamano@suse.com>