Update patch set 1
Patch Set 1:

Hi Sam

no, this code does not need to be merged quickly, this is just to get the cli of ODL if you start a few features in netvirt.
This is not a oldparent/yangtools bump effort.

Patch-set: 1
Reviewer: Gerrit User 6988 <6988@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0