NETVIRT-1068: Upstreaming fixes
Issue-1: When VpnInterface has a secondary adjacency like invisible ip, if tunnel
EP is deleted , VpnInterface Op DS is updated with empty nexthoplist.
During this time , invisible ip goes missing from VpnInterface Op DS due
to MDSAL.syncwrite() being done in updateVpnInterfaceOnTepDelete. Another
fix being done here is , invisible ip FIB entry is updated with empty
nexthop and route is withdrawn/advertized to BGP

Issue-2:  prefix-A part of subnet-A/network-A/VPN-A. The refcount for
prefix-A is 1. When VPN-B is importing from VPN-A , refcount is 2. But
when prefix-B part of VPN-B is deleted , refcount was not decremented.
For self-imported routes , when the prefix is deleted , refcount is
decremented.

Change-Id: I2cabeaa1ae23e356b2d17a527e9bbc6a5b21afd0
Signed-off-by: HANAMANTAGOUD V Kandagal <hanamantagoud.v.kandagal@ericsson.com>