Update patch set 105
Patch Set 105:

We understand that when conntrack support will be get rid off of vpn-id, then this patch will be useless.

May I ask you of which target you are referring to ?
Is [1] planned for both master an nitrogen.

From our perspective, we are commited to merge [0] in both master and nitrogen branch.
It is great that [1] will be available.
In order to avoid not having [0] neither [1] in one branch, we would prefer to have [0] being merged first.

Then a subsequent commit by removing not only the vpn-id list from [1], but also the removal of vpnId from aclservice.

Patch-set: 105