Update patch set 1
Patch Set 1:

Victor,
We considered the following 5 HA configration use cases.
- two nodes (A, B) have same hardware_vtep database schema version
  1) A: support replication mode?B: support replication mode 
  2) A: not support replication mode?B: not support replication mode
- two nodes (A, B) have different hardware_vtep database schema version
  3) A: support replication mode?B: support replication mode 
  4) A: support replication mode?B: not support replication mode
  5) A: not support replication mode?B: not support replication mode
The case methioned in the last comment is case 4.
The cases that two nodes (A, B) have different hardware_vtep database schema version (use case 3,4,5) are rare.
Thus we think it's OK that do not support them currently, and we can implement them as improvement requests if needed in the future.
So the code suggested by Suneelu should be better if we just consider use case 1 and 2.
I'd like to fix it with the code Suneelu suggested.
What's your opinion?

Patch-set: 1