Update patch set 1
Patch Set 1:

agree with kaoru.
when the two nodes A & B  are operating in HA mode we expect them to have the same ovs version.

What is happening here is node A 's version is reflected in its oper datastore.
but its parent P oper datastore node does not have the version.
That is the fix we are suggesting here to copy node A's version to P.
But if node B comes with a different version , then what should we copy to parent P  ? should it be A's or B's.

Patch-set: 1
Reviewer: Gerrit User 7785 <7785@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0