NETVIRT-1074: DNAT traffic is not working if SNAT is Disabled
Problem Description:
======================
FIP to External Network or FIP to FIP traffic is not working if SNAT is
disabled for the router. Since L3_FIB_TABLE (21) -> PSNAT_TABLE(26) flow
is missing on the FIP associated switch.

Solution:
=========
If SNAT is disabled, we need to take care of DNAT (FIP) traffic should
work as expected. By default when router-gw is set with external
network/subnet SNAT is enabled and default NAT flow on Table 21 will be
getting programmed. In case of SNAT is disabled this flow won't get
installed on the required FIP configured switches. Hence FIP traffic is
getting dropped at Table 21. Have added the logic to program the NAT
default flow in Table 21 if SNAT is disabled UC as well.

Change-Id: Ie6c66c9571d5698408317718bfaa61ce677472c4
Signed-off-by: karthikeyan <karthikeyangceb007@gmail.com>