Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Can we have a CSIT job "test-netvirt-current-controller" which continues to validate SNAT UC(with nat-mode set to Controller) going forward post this change

yes, just add a new 'functionality' like gate-snat-controller for the
relevant branches you'd want to be doing this. In this file:

https://github.com/opendaylight/releng-builder/blob/master/jjb/netvirt/netvirt-csit-multi-openstack.yaml#L31-L43

(note that was just the oxygen section. nitro section is below if that
matters).

then add the logic in the patch test yaml, and it should be good:

https://github.com/opendaylight/releng-builder/blob/master/jjb/netvirt/netvirt-patch-test.yaml

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0