Return the tagged ELAN instance
ElanUtils::updateOperationalDataStore() is given an ElanInstance to
update, but because YANG instances are immutable, it has to create a
new one to add the ELAN tag. As a result, calling code has to waiting
for the surrounding MD-SAL transaction to finish, and then retrieve
the updated ElanInstance from the MD-SAL... This patch makes the
update method return the updated ElanInstance, potentially avoiding a
MD-SAL round-trip and enabling more asynchronous operations.

Change-Id: Ia83d65b2d0953c97976d37af0bb4299841925b8e
Signed-off-by: Stephen Kitt <skitt@redhat.com>