Update patch set 4
Patch Set 4:

There three things that gets programmed when a first elan interface comes up on a compute/dpn
1)The other computes vms macs in dmac table
2)elan groups(local & remote)
3)unknown dmac table flow (single flow ) which points to the group

In this patchset step 2 & 3 is moved to eos leader node instead of oper shard leader node.
So there should not be a question of failed to programme a flow because group does not exist.
what is the issue with this ?

Patch-set: 4