netvirt-1115: ecmp configuration on non bgpvpn router is possible again
The issue consists in allowing ecmp configurations for non bgpvpn
routers. Actually, there is a filtering mechanism with BGPVPN that
consists in limiting the number of ECMP entries with the number of
router distinguisher configured on the VPNInstance. In cases, where the
VPNInstance is a router, there is no need to do that.

Change-Id: Ic6ccc5419d6e9228651cbdaf8519a788560358e9
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>