Update patch set 8
Patch Set 8:

(1 comment)

The reason for not using aliveness-monitor.yang:
-1- the yang is located in genius project. Netvirt has no control on that.
-2- currently no one is using it. It can removed/modified as and when needed.

As the content which we are referring is SAME (both in aliveness-monitor and ebgp-bfd). IMO, dont require to change to SPEC. 

If you feel, it is mandatory to update the spec please let me know. I will do the same.

Patch-set: 8
Reviewer: Gerrit User 5401 <5401@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0