Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > Hmm I didnt have any problems but it's clearly a mistake. It should
 > be in dependency management and not in dependencies.

I realized that having a profile here would expose it to consumers, as this pom is imported from other projects. Right?

That would be an option. But for now I decided to remove the profile from this artifacts pom and just add sfc to the standard dependencyManagement section. It worked fine for me.

Other option would be to get rid of the profile all together if we have no specific purpose for it.

Any thoughts?

Patch-set: 1
Label: Verified=0