Update patch set 2
Patch Set 2:

(1 comment)

I ran the patch, it works only if VM1 and VM2 are in different compute nodes. It doesn't work if two VMs are in the same compute node. I don't know whether if this is a separate issue.
The logic to determined 'firstPortOnDpn' is a bit confusing to me

Patch-set: 2
Label: Verified=0