Update patch set 1
Patch Set 1:

Actually my suggested solution doesn't work for scenario that two  VMs in same compute node because it doesn't correctly determine the firstPortOnDpn.

Your initial patch is more correct. Sorry about the confusion. But could you check if the block 491-498 could be removed. I am not sure what it is supposed to do:

                for (AclInterface aclInterface : aclInterfaces) {
                    if (port.getInterfaceId().equals(aclInterface.getInterfaceId())) {
                        continue;
                    }
                    for (AllowedAddressPairs ip : aclInterface.getAllowedAddressPairs()) {
                        updateRemoteAclTableForPort(aclInterface, acl, addOrRemove, ip, aclId, port.getDpId());
                    }
                }

I think the block always hits the 'continue' in the block.

Patch-set: 1