NETVIRT-1142: sfc classifier bind to proper egress interface
SFC classifier binds on the SF egress interface of the last RSP hop to
be able to properly route the traffic to its original destination in
case it is on the same compute as the original source of the traffic.

With the introduction of directional data plane locators for symmetric
paths in SFC, the ingress and egress interfaces of a SF might be
different and the classifier has to take care to bind to the correcti
one depending on the direction of the path, wether it is the reverse
direction or the forward direction of the symmetric path.

Before this patch, the classifier binds to any interface of the SF, as
only one was expected.

Change-Id: I98090ea21d14a62cd19e5ad9065e8c9a90e03c11
JIRA: NETVIRT-1142
Signed-off-by: Jaime Caamaño Ruiz <jcaamano@suse.com>
(cherry picked from commit 5211b68)
Signed-off-by: Jaime Caamaño Ruiz <jcaamano@suse.com>