NETVIRT-430 Ping responder on tenant network and FIP-FIP traffic between vm in same n/w in same compute is not working in stateful SG mode.
Replace the dummy commit with zone 0 with ct_clear action. This clears
the ct state in the current packet and will submit the packet to ct
again. So in cases where ct(table=x, zone=2) action twice in the
pipeline( with same zone) , the second one also reaches the netfilter.

Requires ovs2.9 with kernel module installed.

Change-Id: Iab842c6232476d499a3d8c8602d5b2f57f9e4e1f
Signed-off-by: Aswin Suryanarayanan <asuryana@redhat.com>