NETVIRT-1134 : ModifiedNodeDoesNotExistException bgp/networks in CSIT
This exception was being thrown when a delete was performed on a
non-existent node.
This was because of the data-remaining in the data-stores that were
created during the initial_create-networks on devstack bringup.

This patch aims at proper handling of deletion of those bgp routes that
have no nexthop.
It also prevents withdrawal of extra-routes from bgp when such the router
for the extra-routes is not associated to a bgpvpn

Change-Id: Ia666014c03b3f3454878487d3d42ef756a8bd2b2
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>
Signed-off-by: Sam Hague <shague@redhat.com>