NETVIRT-1146 : Exception IllegalArgumentException due to null subnet-id during extra-route vrf-entry deletion
A bug in one of the checks during extra-route creation was leading
to exceptions in the delete path.
Label-Mapper is a bgpvpn exclusive data-store and should not be populated
for router based interal-vpn adjacencies. Such a population lead to
iRT/eRT code execution during the delete extra route life-cycle for plain
router case, leading to this exception.

Change-Id: I9bbfcd55a56f64793a322c647ad63f60c2a8ca1a
Signed-off-by: eupakir <kiran.n.upadhyaya@ericsson.com>