Update patch set 1
Patch Set 1:

Yes, one other option would be to set the default to false in the yang model. But, that would be a model change, and I'm not exactly sure what the limitations/restrictions are on making yang/model changes and backporting, etc. Stephen, you have any inputs on this? You seemed to frown on this option in the IRC chat in #opendaylight-netvirt earlier today.

Patch-set: 1