Update patch set 10
Patch Set 10:

the latest exception was for a different use-case (extra-route) deletion.
Route-withdrawal from bgp for such extra-routes was being attempted even when they were not part of a bgpvpn. This was due to an incorrect check. I have fixed it in PS10.
The fix is similar to the one that fixed NETVIRT-1146, except that this one is in the extra-route delete path

Patch-set: 10