Update patch set 9
Patch Set 9:

looks like the last tdd trigger didn't create a job so we only have the tdd from the oxygen patch. In that patch was an exception as below. Should that be fixed by this patch here or is that something else being tracked?

2018-03-13T14:37:40,241 | ERROR | qtp2028530521-146 | ContainerResponse                | 38 - com.sun.jersey.jersey-server - 1.17.0 | The RuntimeException could not be mapped to a response, re-throwing to the HTTP container
java.lang.NullPointerException: null
	at org.opendaylight.netvirt.neutronvpn.NeutronvpnManager.dissociateNetworksFromVpn(NeutronvpnManager.java:2394) [359:org.opendaylight.netvirt.neutronvpn-impl:0.6.0.SNAPSHOT]

Patch-set: 9
Label: Verified=0