Update patch set 6
Patch Set 6:

> Build Unstable
 > 
 > https://jenkins.opendaylight.org/releng/job/netvirt-patch-test-current-oxygen/978/
 > : UNSTABLE (skipped)
 > 
 > No problems were identified. If you know why this problem occurred,
 > please add a suitable Cause for it. ( https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-openstack-queens-gate-stateful-oxygen/105/
 > )
 > 
 > Logs: https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/netvirt-patch-test-current-oxygen/978


two failures here. the PNF failure is not expected, but the fix for that was
merged yesterday I think. I think if you rebase that failure will go away
and only the tempest_stop_start failure will be there which has been determined
to be a bug in nova, and ok to see here.

specific link to the csit run from the patch-test job:
  https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-openstack-queens-gate-stateful-oxygen/105/

Patch-set: 6
Label: Code-Review=0