Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Sam,
>  Can we have this PS to merge. Since have cherry-picked this PS to Stable/Oxygen and ran CSIT job there looks like CSIT job is passed all the TCs except one know tempest issue.
> 
> https://jenkins.opendaylight.org/releng/job/netvirt-csit-1node-openstack-queens-gate-stateful-oxygen/112/robot/openstack/05%20tempest/tempest.scenario.test_network_advanced_server_ops.TestNetworkAdvancedServerOps.test_server_connectivity_stop_start/
> 
> 
> Thanks,
> Karthikeyan.

Since this is not for a blocker bug, I think we should hold off on
asking to get this merged until the Oxygen release is finished and
the branch is unlocked. Of course, if this is deemed a serious enough
bug, we can still lobby to get it in.

JamO

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0