Update patch set 3
Patch Set 3:

Q: "Can you please educate me where do you ever do cleanupDpnOnVpn for the external-network VPN if not here?"

A: The cleanup DPN for external subnet VPN will be done when the external subnet is deleted. Since these PNF flows are added to the relevant dpns when the external subnet is created, it is logical to remove the flows when the external subnet is removed.
In this case all the Fib entries belong to the external subnet will be removed. 

The method fibmanager.cleanupDpnOnVpn is invoked only when an interface in a VPN is removed.

Patch-set: 3