Update patch set 3
Patch Set 3:

Hi Vinh,

Reviewed the code and this looks to be a detour for cleanupDpnOnVpn.

Can I kindly request you that we fix the rootcause where conntrack based NATing puts the router-gateway-ip into vpn-to-dpn-list similar to how Controller-based-NATing puts in today so that existing cleanUpDpnOnVpn works. 

Please note that CleanUpDpnOnVPn is vpnEngine driven and has no coupling with NAT use-cases at all.  And so making changes in that area for NAT cases would be better to avoid.

Patch-set: 3