Update patch set 2
Patch Set 2: Code-Review-2

Hi Sam,

You can ignore that patch for now. I will fix the verify jobs up next week. There's 2 issues with the jobs that need to be resolved to get them passing and we have patches for both that need to be reviewed and a new global-jjb released.

Regards,
Thanh

Patch-set: 2
Label: Code-Review=-2