Update patch set 3
Patch Set 3:

> test-genius-netvirt

use test-netvirt -current - without the space, as if I enter the keyword it will trigger since the keyword search is a simple grep - if the trigger is anywhere in the comment it will be found.

Each test trigger is of the form test-<project>-<functionality>, so test-genius-netvirt is to be used from the genius patches.

Patch-set: 3