Clean up VPN data on VPN delete.
The prefix-to-interface, l3-nexthop datastores will not have any data for
a vpn, until a vpn-interface is up and running in the network.
Under such conditions VPN deletion causes
ModifiedNodeDoesNotExistException on trying to delete such datastores
which causes entire transaction which handles clean-up to fail, leave
stale-data in the system.

This further causes new VPN-create with same RDs to fail. This review
fixes the issue.

Change-Id: I3011d81907eb32693de3079bd118cfaaf1b9dfcc
Signed-off-by: Deepthi V V <deepthi.v.v@ericsson.com>