NETVIRT-1171:IPv4 entries appearing in FIB though not associated to Router
Problem Description:
====================
IPv4 entries are appearing in FIB when subnet is not associated to router with
dualstack VM

Solution Description:
=====================
Dual Stack VM has both IPv4 and IPv6 Address, IPv6 or IPv4 subnet only
associated to router means it should show only IPv6 or IPv4 address based
on the subnet address family. To do this behaviour we need to create the
VPN-Interface adjacency based on the subnet associated to router. Have
done the code changes accordingly in VPN-Interface adjacency.

Change-Id: Id1a4712dc1b8be0713fe7dcb526555c5e5021e39
Signed-off-by: Karthikeyan Krishnan <karthikeyangceb007@gmail.com>