Update patch set 1
Patch Set 1: Code-Review-1

> > Ideally the RA Src IP must correspond to the LLA of the subnet
 > GW.
 > > With this fix, it will allow RA from ANY host in that subnet. Is
 > > there any possibility of closing this gap without additional
 > > orchestration?
 > 
 > I didn't find any straight forward way to get LLA of subnet GW. But
 > it can be framed from the IPv6 subnet's router interface
 > MAC-Address which is a bit long process, hence having matching on
 > "fe80::/10" for now.

I agree with Asvin that we should not open up the entire fe80::/10 as it could pose security issues. The ACL rule should be programmed in such a way that we only allow RA from the LLA of the external ipv6 subnet-gateway-ip. Now in-order to do that the orchestration would be that we expect user to explicitly specify the gateway-ip (as LLA) during subnet-creation.

Patch-set: 1
Label: Code-Review=-1