Update patch set 4
Patch Set 4:

> (1 comment)

Hi Karthikeyan,

The Conntrack SNAT  external router g/w is set is decoupled from flow programming so this method will not be invoked. All the flow programming happens in response to centralizedswitch change event. If you think in case of controller SNAT this method will be invoked twice I can add the check to do it only conntrack mode. But the CentralizedSwitchChangeListener is now always failing to delete the flows, which needs to be addressed I think.

Patch-set: 4