Prevent new code from using the DataBroker
This introduces a new Checkstyle check to fail the build if code
*calls* the newReadWriteTransaction or newWriteOnlyTransaction methods
in DataBroker. I haven’t found a nice way to make this check
conditional, so for now it’s added via a new parent module,
managed-tx-parent, which is used instead of binding-parent in all
modules which currently support it.

This patch is self-contained, but the plan is to move the Checkstyle
implementation to infrautils and the Checkstyle configuration to
Genius or MD-SAL, alongside ManagedNewTransactionRunner. Once all the
existing code has been converted to use only
ManagedNewTransactionRunner, we’ll be able to apply the configuration
to binding-parent and do away with managed-tx-parent.

Change-Id: I27ab7f0744477606bae04bff31a2148134038bb8
Signed-off-by: Stephen Kitt <skitt@redhat.com>