Update patch set 1
Patch Set 1:

> > Patch Set 1: Code-Review-1
 > >
 > > (1 comment)
 > 
 > Hi Sridhar,
 > 
 > I am a bit confused as I do not follow your comment. While it is
 > true that DHCP port is a "real tap interface", there is no
 > expectation of connectivity to this interface via L3 VPN service.
 > So, this entry should NOT be in the FIB, independent of whether
 > router is associated with the subnet or not.
 > 
 > There will be DHCP port created, one per ELAN. So the connectivity
 > model would be based on ELAN and never VPN.
 > 
 > Can you please clarify?
 > Thanks,
 > Ashvin

Hello Ashvin,

Lets say, I want to ping the dhcp-interface from a different network or from some external server (in a "FLAT/VLAN" based provider network scenario). 

We should be able to support this use-case. Now, in-order to do this, we would require an entry in FIB (which works today).

AFAIU, this patch removes this FIB entry for neutron dhcp-port un-conditionally. Hope this helps.

Thanks,
--Sridhar.

Patch-set: 1