<WIP>VMs on tennant vlan network not getting an IP when spwaned after the networks are added to the router.
The router port elandpnlist didn't have a state. So when the first port
is router port in the list, the elan flows fails.

Check is added to skip router port while checking if a port is the first
one in the list.

NETVIRT-1208 An Instance doesn't get an IP after when a network is added
to a router

Change-Id: I25ea20759eefc3043f1616e7907b0c6916fe90e5
Signed-off-by: Aswin Suryanarayanan <asuryana@redhat.com>