Update L3_GW_MAC table to match on ip/ipv6 for external network
Currently, for ingress traffic (i.e., external network), L3_GW_MAC(19)
table only includes a match for dl_dst/mac-address. When an external
network that has both IPv4 and IPv6 subnets is associated to a neutron
router, the router-gw-mac would be the same for both the subnets.
Ideally, L3_GW_MAC table should check the type of subnet and then
update the metadata to either vpnId of ext-ipv4-subnet or ext-ipv6-subnet
before sending the packet to FIB table. This patch addresses this issue.

Issue: NETVIRT-1220
Change-Id: Id139fff3b6a9ee25bc9c9bda3d7521f7d2ff95fb
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>