Router internal address not reachable
The router internal address was not reachable since ovs is marking the
packet as -trk. Now -trk packets will be resubmitted to ct to have
second lookup.

This is a workaround until ovs bug is fixed. Now in table 239 ct_clear
is called for both tracked and untracked packet.

NETVIRT-1064 No connectivity to a router's internal address

Change-Id: Id5b2500822d6cc37c8ee82feae0aa9ee75cb0426
Signed-off-by: Aswin Suryanarayanan <asuryana@redhat.com>