Update patch set 1
Patch Set 1:

> what is this stuff?

The test tools Vishal and I have been working on. This particular patch is the csit robot files parsing piece. Feed in an output.xml.gz file and it will extract all the debug files per suite and test so that you have all the model and flow dumps as individual files. From there the files can be further processed but other tools in this package and they will add stuff like metadata and table names.

Patch-set: 1